Fix passive/active mode commands
Not sure why this ever worked, as it seems most FTP clients expect you to open the port/connect to the client port immediately after issuing the PORT/PASV commands. I think it's probably related to some sort of race condition.

Now, the socket is created as soon as PASV/PORT is called.